fix(#44): move io_profile.py from core/ to contracts/
Pure data types (StrEnum + frozen dataclass) with no kernel deps.
- Move core/io_profile.py → contracts/io_profile.py
- Update 5 caller files (28 import sites total via replace_all)
