+++ src/nexus/contracts/io_profile.py
