+++ tests/unit/core/test_io_profile.py
@@ -12,12 +12,12 @@ class TestIOProfileEnum:
     """Test IOProfile enum values and behavior."""
 
     def test_has_six_values(self) -> None:
-        from nexus.core.io_profile import IOProfile
+        from nexus.contracts.io_profile import IOProfile
 
         assert len(IOProfile) == 6
 
     def test_string_values(self) -> None:
-        from nexus.core.io_profile import IOProfile
+        from nexus.contracts.io_profile import IOProfile
 
         assert IOProfile.FAST_READ == "fast_read"
         assert IOProfile.FAST_WRITE == "fast_write"
@@ -27,25 +27,25 @@ class TestIOProfileEnum:
         assert IOProfile.ARCHIVE == "archive"
 
     def test_str_conversion(self) -> None:
-        from nexus.core.io_profile import IOProfile
+        from nexus.contracts.io_profile import IOProfile
 
         assert str(IOProfile.FAST_READ) == "fast_read"
         assert str(IOProfile.BALANCED) == "balanced"
 
     def test_from_string(self) -> None:
-        from nexus.core.io_profile import IOProfile
+        from nexus.contracts.io_profile import IOProfile
 
         assert IOProfile("fast_read") is IOProfile.FAST_READ
         assert IOProfile("balanced") is IOProfile.BALANCED
 
     def test_invalid_value_raises(self) -> None:
-        from nexus.core.io_profile import IOProfile
+        from nexus.contracts.io_profile import IOProfile
 
         with pytest.raises(ValueError):
             IOProfile("nonexistent")
 
     def test_config_method_returns_profile_config(self) -> None:
-        from nexus.core.io_profile import IOProfile, IOProfileConfig
+        from nexus.contracts.io_profile import IOProfile, IOProfileConfig
 
         cfg = IOProfile.BALANCED.config()
         assert isinstance(cfg, IOProfileConfig)
@@ -55,7 +55,7 @@ class TestIOProfileConfig:
     """Test IOProfileConfig frozen dataclass."""
 
     def test_frozen_dataclass(self) -> None:
-        from nexus.core.io_profile import IOProfileConfig
+        from nexus.contracts.io_profile import IOProfileConfig
 
         cfg = IOProfileConfig(
             readahead_enabled=True,
@@ -72,7 +72,7 @@ class TestIOProfileConfig:
             cfg.readahead_enabled = False  # type: ignore[misc]
 
     def test_all_fields_accessible(self) -> None:
-        from nexus.core.io_profile import IOProfileConfig
+        from nexus.contracts.io_profile import IOProfileConfig
 
         cfg = IOProfileConfig(
             readahead_enabled=True,
@@ -104,7 +104,7 @@ class TestProfileConfigMapping:
         ["fast_read", "fast_write", "edit", "append_only", "balanced", "archive"],
     )
     def test_each_profile_has_config(self, profile_name: str) -> None:
-        from nexus.core.io_profile import IOProfile
+        from nexus.contracts.io_profile import IOProfile
 
         profile = IOProfile(profile_name)
         cfg = profile.config()
@@ -124,7 +124,7 @@ class TestProfileConfigMapping:
     def test_readahead_enabled_per_profile(
         self, profile_name: str, expected_readahead: bool
     ) -> None:
-        from nexus.core.io_profile import IOProfile
+        from nexus.contracts.io_profile import IOProfile
 
         cfg = IOProfile(profile_name).config()
         assert cfg.readahead_enabled is expected_readahead
@@ -141,27 +141,27 @@ class TestProfileConfigMapping:
         ],
     )
     def test_cache_priority_per_profile(self, profile_name: str, expected_priority: int) -> None:
-        from nexus.core.io_profile import IOProfile
+        from nexus.contracts.io_profile import IOProfile
 
         cfg = IOProfile(profile_name).config()
         assert cfg.cache_priority == expected_priority
 
     def test_fast_read_has_highest_readahead_window(self) -> None:
-        from nexus.core.io_profile import IOProfile
+        from nexus.contracts.io_profile import IOProfile
 
         fast_read = IOProfile.FAST_READ.config()
         balanced = IOProfile.BALANCED.config()
         assert fast_read.readahead_max_window >= balanced.readahead_max_window
 
     def test_fast_write_has_largest_write_buffer(self) -> None:
-        from nexus.core.io_profile import IOProfile
+        from nexus.contracts.io_profile import IOProfile
 
         fast_write = IOProfile.FAST_WRITE.config()
         balanced = IOProfile.BALANCED.config()
         assert fast_write.write_buffer_max_size >= balanced.write_buffer_max_size
 
     def test_archive_all_zeroes(self) -> None:
-        from nexus.core.io_profile import IOProfile
+        from nexus.contracts.io_profile import IOProfile
 
         cfg = IOProfile.ARCHIVE.config()
         assert cfg.readahead_enabled is False
@@ -173,13 +173,13 @@ class TestProfileConfigMapping:
         assert cfg.write_buffer_max_size == 0
 
     def test_edit_sync_mode_enabled(self) -> None:
-        from nexus.core.io_profile import IOProfile
+        from nexus.contracts.io_profile import IOProfile
 
         cfg = IOProfile.EDIT.config()
         assert cfg.write_buffer_sync_mode is True
 
     def test_fast_read_prefetch_on_open(self) -> None:
-        from nexus.core.io_profile import IOProfile
+        from nexus.contracts.io_profile import IOProfile
 
         cfg = IOProfile.FAST_READ.config()
         assert cfg.readahead_prefetch_on_open is True
@@ -189,7 +189,7 @@ class TestReadaheadConfigDerivation:
     """Test IOProfile → ReadaheadConfig conversion."""
 
     def test_from_io_profile_fast_read(self) -> None:
-        from nexus.core.io_profile import IOProfile
+        from nexus.contracts.io_profile import IOProfile
         from nexus.fuse.readahead import ReadaheadConfig
 
         cfg = ReadaheadConfig.from_io_profile(IOProfile.FAST_READ)
@@ -200,14 +200,14 @@ class TestReadaheadConfigDerivation:
         assert cfg.prefetch_on_open is True
 
     def test_from_io_profile_fast_write_disabled(self) -> None:
-        from nexus.core.io_profile import IOProfile
+        from nexus.contracts.io_profile import IOProfile
         from nexus.fuse.readahead import ReadaheadConfig
 
         cfg = ReadaheadConfig.from_io_profile(IOProfile.FAST_WRITE)
         assert cfg.enabled is False
 
     def test_from_io_profile_balanced(self) -> None:
-        from nexus.core.io_profile import IOProfile
+        from nexus.contracts.io_profile import IOProfile
         from nexus.fuse.readahead import ReadaheadConfig
 
         cfg = ReadaheadConfig.from_io_profile(IOProfile.BALANCED)
@@ -221,7 +221,7 @@ class TestReadaheadConfigDerivation:
         ["fast_read", "fast_write", "edit", "append_only", "balanced", "archive"],
     )
     def test_all_profiles_produce_valid_config(self, profile_name: str) -> None:
-        from nexus.core.io_profile import IOProfile
+        from nexus.contracts.io_profile import IOProfile
         from nexus.fuse.readahead import ReadaheadConfig
 
         cfg = ReadaheadConfig.from_io_profile(IOProfile(profile_name))
@@ -232,18 +232,18 @@ class TestCachePriorityMapping:
     """Test IOProfile → cache priority values."""
 
     def test_priority_range(self) -> None:
-        from nexus.core.io_profile import IOProfile
+        from nexus.contracts.io_profile import IOProfile
 
         for profile in IOProfile:
             priority = profile.config().cache_priority
             assert 0 <= priority <= 3, f"{profile}: priority {priority} out of range"
 
     def test_fast_read_highest_priority(self) -> None:
-        from nexus.core.io_profile import IOProfile
+        from nexus.contracts.io_profile import IOProfile
 
         assert IOProfile.FAST_READ.config().cache_priority == 3
 
     def test_archive_lowest_priority(self) -> None:
-        from nexus.core.io_profile import IOProfile
+        from nexus.contracts.io_profile import IOProfile
 
         assert IOProfile.ARCHIVE.config().cache_priority == 0
